+++ server/src/dtsTypes/standardTypes/nodeTypes/cpus/node.ts
@@ -53,11 +53,13 @@ export function getCpusNodeType() {
   };
 
   const sizeCellsAdditionalTypeCheck = sizeCellsProp?.additionalTypeCheck;
-  sizeCellsProp!.additionalTypeCheck = (property) => {
-    const issues = sizeCellsAdditionalTypeCheck?.(property) ?? [];
+  sizeCellsProp!.additionalTypeCheck = (property, macros) => {
+    const issues = sizeCellsAdditionalTypeCheck?.(property, macros) ?? [];
 
     const node = property.parent.getProperty(property.name);
-    const nodeValue = node ? getU32ValueFromProperty(node, 0, 0) : undefined;
+    const nodeValue = node
+      ? getU32ValueFromProperty(node, 0, 0, macros)
+      : undefined;
 
     if (nodeValue !== 0) {
       issues.push(