+++ server/src/dtsTypes/standardTypes/reg.ts
@@ -16,7 +16,11 @@
 
 import { BindingPropertyType } from "../../types/index";
 import { FileDiagnostic, StandardTypeIssue } from "../../types";
-import { genStandardTypeDiagnostic } from "../../helpers";
+import {
+  compareWords,
+  createTokenIndex,
+  genStandardTypeDiagnostic,
+} from "../../helpers";
 import { PropertyNodeType } from "../types";
 import {
   flatNumberValues,
@@ -24,6 +28,8 @@ import {
   getU32ValueFromProperty,
 } from "./helpers";
 import { DiagnosticSeverity } from "vscode-languageserver";
+import { ASTBase } from "../../ast/base";
+import { ArrayValues } from "../../ast/dtc/values/arrayValue";
 
 export default () => {
   const prop = new PropertyNodeType<number>(
@@ -34,7 +40,7 @@ export default () => {
     },
     undefined,
     undefined,
-    (property) => {
+    (property, macros) => {
       const issues: FileDiagnostic[] = [];
 
       const values = flatNumberValues(property.ast.values);
@@ -42,32 +48,24 @@ export default () => {
         return [];
       }
 
-      const sizeCellProperty =
-        property.parent.parent?.getProperty("#size-cells");
-      const addressCellProperty =
-        property.parent.parent?.getProperty("#address-cells");
-
-      const sizeCell = sizeCellProperty
-        ? getU32ValueFromProperty(sizeCellProperty, 0, 0) ?? 1
-        : 1;
-      const addressCell = addressCellProperty
-        ? getU32ValueFromProperty(addressCellProperty, 0, 0) ?? 2
-        : 2;
+      const parentSizeCell = property.parent.parentSizeCells(macros);
+      const parentAddressCell = property.parent.parentAddressCells(macros);
 
       prop.typeExample = `<${[
-        ...Array.from({ length: addressCell }, () => "address"),
-        ...Array.from({ length: sizeCell }, () => "cell"),
+        ...Array.from({ length: parentAddressCell }, () => "address"),
+        ...Array.from({ length: parentSizeCell }, () => "cell"),
       ].join(" ")}>`;
 
       if (
         values.length === 0 ||
-        values.length % (sizeCell + addressCell) !== 0
+        values.length % (parentSizeCell + parentAddressCell) !== 0
       ) {
         issues.push(
           genStandardTypeDiagnostic(
             StandardTypeIssue.CELL_MISS_MATCH,
             values.at(
-              values.length - (values.length % (sizeCell + addressCell))
+              values.length -
+                (values.length % (parentSizeCell + parentAddressCell))
             ) ?? property.ast,
             DiagnosticSeverity.Error,
             [],
@@ -78,45 +76,35 @@ export default () => {
         return issues;
       }
 
-      for (let i = 0; i < values.length; i += sizeCell + addressCell) {
-        const buffer = new ArrayBuffer(addressCell * 4);
-        const view = new DataView(buffer);
+      for (
+        let i = 0;
+        i < values.length;
+        i += parentSizeCell + parentAddressCell
+      ) {
+        const refAddress = values
+          .slice(0, parentAddressCell)
+          .map((_, i) => getU32ValueFromProperty(property, 0, i, macros) ?? 0);
 
-        values
-          .slice(0, addressCell)
-          .map((_, i) => getU32ValueFromProperty(property, 0, i) ?? 0)
-          .forEach((c, i) => {
-            view.setUint32(i * 4, c);
-          });
-
-        // TODO consider adding check and warn if property.parent.address?.length > 1 && property.parent.address?.length !== addressCell
-
-        if (property.parent.address && property.parent.address?.length > 1) {
-          property.parent.address?.forEach((a, i) => {
-            if (view.getUint32(i * 4) !== a) {
-              issues.push(
-                genStandardTypeDiagnostic(
-                  StandardTypeIssue.MISMATCH_NODE_ADDRESS_REF_FIRST_VALUE,
-                  property.ast,
-                  DiagnosticSeverity.Error,
-                  [],
-                  [],
-                  [property.name]
-                )
-              );
-            }
-          });
-        } else if (
-          property.parent.address?.length === 1 &&
-          ((addressCell === 2 &&
-            view.getBigUint64(0) !== BigInt(property.parent.address[0])) ||
-            (addressCell === 1 &&
-              view.getUint32(0) !== property.parent.address[0]))
+        if (
+          property.parent.address &&
+          compareWords(property.parent.address, refAddress) !== 0
         ) {
+          const startValue = (
+            property.ast.values?.values.at(0)?.value as ArrayValues | undefined
+          )?.values?.at(0);
+          const endValue = (
+            property.ast.values?.values.at(0)?.value as ArrayValues | undefined
+          )?.values?.at(parentAddressCell - 1);
+          const addressValues =
+            startValue && endValue
+              ? new ASTBase(
+                  createTokenIndex(startValue.firstToken, endValue.lastToken)
+                )
+              : undefined;
           issues.push(
             genStandardTypeDiagnostic(
-              StandardTypeIssue.MISMATCH_NODE_ADDRESS_REF_FIRST_VALUE,
-              property.ast,
+              StandardTypeIssue.MISMATCH_NODE_ADDRESS_REF_ADDRESS_VALUE,
+              addressValues ?? property.ast.values ?? property.ast,
               DiagnosticSeverity.Error,
               [],
               [],