+++ server/src/server.ts
@@ -275,7 +275,7 @@ connection.onInitialize((params: InitializeParams) => {
   // The workspace folder this server is operating on
   workspaceFolder = params.workspaceFolders;
   connection.console.log(
-    `[Server(${process.pid}) ${workspaceFolder?.[0].uri} Version 0.3.1 ] Started and initialize received`
+    `[Server(${process.pid}) ${workspaceFolder?.[0].uri} Version 0.3.2 ] Started and initialize received`
   );
 
   const capabilities = params.capabilities;
@@ -810,7 +810,8 @@ const updateActiveContext = async (id: ContextId, force = false) => {
     if (newContext) {
       contexMeta(newContext).then(async (meta) => {
         const fileTree = await newContext.getFileTree();
-        if (newContext !== activeContext) return;
+        if (newContext !== activeContext && !contextAware.includes(newContext))
+          return;
         connection.sendNotification("devicetree/newActiveContext", {
           ctxNames: newContext.ctxNames.map((c) => c.toString()),
           id: newContext.id,